+++ server/auth/__mocks__/generators.ts
@@ -10,20 +10,24 @@ import { jest } from "@jest/globals";
 
 export const compare = jest
 	.fn<typeof _compare>()
-	.mockImplementation((a: unknown, b: unknown) => Promise.resolve(a === b));
+	.mockImplementation((a: string, b: string) => Promise.resolve(a === b));
+
+export const DEFAULT_MOCK_SECURE_TOKEN = "NOT_SECURE_TOKEN" as TOTPSeed;
 
 export const generateSecureToken = jest
 	.fn<typeof _generateSecureToken>()
-	.mockReturnValue("NOT_SECURE_TOKEN" as TOTPSeed);
+	.mockReturnValue(DEFAULT_MOCK_SECURE_TOKEN);
 
-export const generateSalt = jest
-	.fn<typeof _generateSalt>()
-	.mockResolvedValue("INSECURE_SALT" as Salt);
+export const DEFAULT_MOCK_SALT = "INSECURE_SALT" as Salt;
 
-export const generateHash = jest
-	.fn<typeof _generateHash>()
-	.mockResolvedValue("INSECURE_HASH" as Hash);
+export const generateSalt = jest.fn<typeof _generateSalt>().mockResolvedValue(DEFAULT_MOCK_SALT);
+
+export const DEFAULT_MOCK_HASH = "INSECURE_HASH" as Hash;
+
+export const generateHash = jest.fn<typeof _generateHash>().mockResolvedValue(DEFAULT_MOCK_HASH);
+
+export const DEFAULT_MOCK_AES_CIPHER_KEY = "INSECURE_CIPHER_KEY" as AESCipherKey;
 
 export const generateAESCipherKey = jest
 	.fn<typeof _generateAESCipherKey>()
-	.mockResolvedValue("INSECURE_CIPHER_KEY" as AESCipherKey);
+	.mockResolvedValue(DEFAULT_MOCK_AES_CIPHER_KEY);