+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Our Express environment now sends the `Access-Control-Allow-Headers` header with each request, as our Vercel environment does.
+
 ## [0.18.1] - 2023-03-24
 ### Added
 - A management UI for viewing database contents. (These should be encrypted client-side, so there's no user data to see here. I just don't wanna have to write a query string every time I need to read other parts of the database.)
@@ -460,6 +464,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial commit
 
+[Unreleased]: https://github.com/RecordedFinance/recorded-finance/compare/v0.18.1...HEAD
 [0.18.1]: https://github.com/RecordedFinance/recorded-finance/compare/v0.18.0...v0.18.1
 [0.18.0]: https://github.com/RecordedFinance/recorded-finance/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/RecordedFinance/recorded-finance/compare/v0.16.8...v0.17.0